+++ CS575/Project_7/Project_7B_CL_Alferman.cpp
@@ -282,9 +282,9 @@ int main ( int argc, char *argv[ ] )
 
   // Calculate computational speed
 
-  float SpeedSerial = Size / (time1-time0) / 1000;
-  float SpeedOpenMP = Size / (time2-time1) / 1000;
-  float SpeedCL = Size / (time5-time2) / 1000;
+  float SpeedSerial = (double) Size / (time1-time0) / 1000.;
+  float SpeedOpenMP = (double) Size / (time2-time1) / 1000.;
+  float SpeedCL = (double) Size / (time5-time2) / 1000.;
 
   // Print out the speeds
